Add Animation::LoadXML. Use xml animation in Sample 30.
