+++ bin/Data/Animations/LightAnimation.xml
@@ -1,0 +1,16 @@
+<animation length="4.0">
+    <variant name="@/Position" type="Vector3" interpolation="spline" tension="0.7">
+        <keyframe time="0.0" value="-30.0 5.0 -30.0"/>
+        <keyframe time="1.0" value=" 30.0 5.0 -30.0"/>
+        <keyframe time="2.0" value=" 30.0 5.0  30.0"/>
+        <keyframe time="3.0" value="-30.0 5.0  30.0"/>
+        <keyframe time="4.0" value="-30.0 5.0 -30.0"/>
+    </variant>
+    <variant name="@Light/Color" type="Color">
+        <keyframe time="0.0" value="1 1 1 1"/>
+        <keyframe time="1.0" value="1 0 0 1"/>
+        <keyframe time="2.0" value="1 1 0 1"/>
+        <keyframe time="3.0" value="0 1 0 1"/>
+        <keyframe time="4.0" value="1 1 1 1"/>
+    </variant>
+</animation>
